# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/me/account": {
      "get": {
        "operationId": "AccountController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "phone": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "locale": {
                      "type": "string",
                      "enum": [
                        "ru",
                        "hy",
                        "en"
                      ]
                    },
                    "twoFactorEnabled": {
                      "type": "boolean"
                    },
                    "sessionsRevokedAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "deleteRequestedAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "deletionAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "profile": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "gender": {
                              "type": "string"
                            },
                            "birthday": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "district": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "photoUrl": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "bigFont": {
                              "type": "boolean"
                            },
                            "timeFormat": {
                              "type": "string"
                            },
                            "consentAt": {
                              "type": [
                                "string",
                                "null"
                              ]
                            }
                          },
                          "required": [
                            "gender",
                            "birthday",
                            "district",
                            "photoUrl",
                            "bigFont",
                            "timeFormat",
                            "consentAt"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "version": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "phone",
                    "locale",
                    "twoFactorEnabled",
                    "sessionsRevokedAt",
                    "deleteRequestedAt",
                    "deletionAt",
                    "profile",
                    "version"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Профиль, язык, крупный шрифт, 2FA, удаление (F-00-124, F-15-158)",
        "tags": [
          "account"
        ]
      },
      "patch": {
        "operationId": "AccountController_patch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "locale": {
                    "type": "string",
                    "enum": [
                      "ru",
                      "hy",
                      "en"
                    ]
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "female",
                      "male",
                      "unknown"
                    ]
                  },
                  "birthday": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "district": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 40
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "photoUrl": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 500
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "bigFont": {
                    "type": "boolean"
                  },
                  "timeFormat": {
                    "type": "string",
                    "enum": [
                      "24h",
                      "12h"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "phone": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "locale": {
                      "type": "string",
                      "enum": [
                        "ru",
                        "hy",
                        "en"
                      ]
                    },
                    "twoFactorEnabled": {
                      "type": "boolean"
                    },
                    "sessionsRevokedAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "deleteRequestedAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "deletionAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "profile": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "gender": {
                              "type": "string"
                            },
                            "birthday": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "district": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "photoUrl": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "bigFont": {
                              "type": "boolean"
                            },
                            "timeFormat": {
                              "type": "string"
                            },
                            "consentAt": {
                              "type": [
                                "string",
                                "null"
                              ]
                            }
                          },
                          "required": [
                            "gender",
                            "birthday",
                            "district",
                            "photoUrl",
                            "bigFont",
                            "timeFormat",
                            "consentAt"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "version": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "phone",
                    "locale",
                    "twoFactorEnabled",
                    "sessionsRevokedAt",
                    "deleteRequestedAt",
                    "deletionAt",
                    "profile",
                    "version"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Изменить профиль. If-Match: version — чужая правка между чтением и записью → 409",
        "tags": [
          "account"
        ]
      }
    },
    "/v1/me/account/delete": {
      "post": {
        "operationId": "AccountController_requestDeletion",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "phone": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "locale": {
                      "type": "string",
                      "enum": [
                        "ru",
                        "hy",
                        "en"
                      ]
                    },
                    "twoFactorEnabled": {
                      "type": "boolean"
                    },
                    "sessionsRevokedAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "deleteRequestedAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "deletionAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "profile": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "gender": {
                              "type": "string"
                            },
                            "birthday": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "district": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "photoUrl": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "bigFont": {
                              "type": "boolean"
                            },
                            "timeFormat": {
                              "type": "string"
                            },
                            "consentAt": {
                              "type": [
                                "string",
                                "null"
                              ]
                            }
                          },
                          "required": [
                            "gender",
                            "birthday",
                            "district",
                            "photoUrl",
                            "bigFont",
                            "timeFormat",
                            "consentAt"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "version": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "phone",
                    "locale",
                    "twoFactorEnabled",
                    "sessionsRevokedAt",
                    "deleteRequestedAt",
                    "deletionAt",
                    "profile",
                    "version"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Удалить аккаунт через 25 дней (можно отменить)",
        "tags": [
          "account"
        ]
      }
    },
    "/v1/me/account/delete/cancel": {
      "post": {
        "operationId": "AccountController_cancelDeletion",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "phone": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "locale": {
                      "type": "string",
                      "enum": [
                        "ru",
                        "hy",
                        "en"
                      ]
                    },
                    "twoFactorEnabled": {
                      "type": "boolean"
                    },
                    "sessionsRevokedAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "deleteRequestedAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "deletionAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "profile": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "gender": {
                              "type": "string"
                            },
                            "birthday": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "district": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "photoUrl": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "bigFont": {
                              "type": "boolean"
                            },
                            "timeFormat": {
                              "type": "string"
                            },
                            "consentAt": {
                              "type": [
                                "string",
                                "null"
                              ]
                            }
                          },
                          "required": [
                            "gender",
                            "birthday",
                            "district",
                            "photoUrl",
                            "bigFont",
                            "timeFormat",
                            "consentAt"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "version": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "phone",
                    "locale",
                    "twoFactorEnabled",
                    "sessionsRevokedAt",
                    "deleteRequestedAt",
                    "deletionAt",
                    "profile",
                    "version"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Отменить удаление аккаунта",
        "tags": [
          "account"
        ]
      }
    },
    "/v1/me/account/phone/code": {
      "post": {
        "operationId": "AccountController_phoneCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 24
                  },
                  "channel": {
                    "default": "telegram",
                    "type": "string",
                    "enum": [
                      "telegram",
                      "whatsapp",
                      "sms"
                    ]
                  }
                },
                "required": [
                  "phone"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "challengeId": {
                      "type": "string"
                    },
                    "resendAfter": {
                      "type": "number"
                    },
                    "expiresIn": {
                      "type": "number"
                    },
                    "channel": {
                      "type": "string",
                      "enum": [
                        "telegram",
                        "whatsapp",
                        "sms"
                      ]
                    }
                  },
                  "required": [
                    "challengeId",
                    "resendAfter",
                    "expiresIn",
                    "channel"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Смена номера: код на новый номер (F-15-149)",
        "tags": [
          "account"
        ]
      }
    },
    "/v1/me/account/phone/confirm": {
      "post": {
        "operationId": "AccountController_phoneConfirm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 24
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{4}$"
                  }
                },
                "required": [
                  "phone",
                  "code"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "phone": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "locale": {
                      "type": "string",
                      "enum": [
                        "ru",
                        "hy",
                        "en"
                      ]
                    },
                    "twoFactorEnabled": {
                      "type": "boolean"
                    },
                    "sessionsRevokedAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "deleteRequestedAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "deletionAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "profile": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "gender": {
                              "type": "string"
                            },
                            "birthday": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "district": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "photoUrl": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "bigFont": {
                              "type": "boolean"
                            },
                            "timeFormat": {
                              "type": "string"
                            },
                            "consentAt": {
                              "type": [
                                "string",
                                "null"
                              ]
                            }
                          },
                          "required": [
                            "gender",
                            "birthday",
                            "district",
                            "photoUrl",
                            "bigFont",
                            "timeFormat",
                            "consentAt"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "version": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "phone",
                    "locale",
                    "twoFactorEnabled",
                    "sessionsRevokedAt",
                    "deleteRequestedAt",
                    "deletionAt",
                    "profile",
                    "version"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Смена номера: подтвердить код",
        "tags": [
          "account"
        ]
      }
    },
...
    "/v1/me/consent": {
      "get": {
        "operationId": "AccountController_consent",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Принято ли пользовательское соглашение (F-14-008)",
        "tags": [
          "account"
        ]
      },
      "post": {
        "operationId": "AccountController_acceptConsent",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Принять соглашение (человек вошёл как бизнес и открыл приложение клиента)",
        "tags": [
          "account"
        ]
      }
    },
...
    "/v1/me/sessions/{id}": {
      "delete": {
        "operationId": "AccountController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Закрыть вход на одном устройстве",
        "tags": [
          "account"
        ]
      }
    },
...
  },
  "info": {
    "title": "BookTime API",
    "description": "Сервер для booking-platform. Ошибки — { code, message } (docs/backend/02-api.md §0).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}